fix(server): http dispatch honors turn.start bootstrap
POST /api/orchestration/dispatch passed commands straight to the
orchestration engine, skipping the bootstrap steps the WebSocket path
runs for a thread.turn.start carrying `bootstrap`. An HTTP turn start for
a not-yet-existing thread therefore failed with "Thread does not exist"
even though the contract advertises bootstrap on that endpoint.

The bootstrap flow (thread.create, worktree preparation, setup script,
rollback via thread.delete on failure) now lives in a shared
TurnStartBootstrap service that both the WebSocket RPC and the HTTP
dispatch route use. The ws path keeps its client-origin stamping and
command gating; HTTP gains the bootstrap branch. Tests cover HTTP
bootstrap success and rollback alongside the existing ws cases.
